+++ versions.tf
@@ -19,11 +19,11 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = "< 7.0.0"
+      version = ">= 6.38.0, < 7.0.0"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = "< 7.0.0"
+      version = ">= 6.38.0, < 7.0.0"
     }
     helm = {
       source  = "hashicorp/helm"